+++ app/Http/Controllers/api/v1/ItemController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\api\v1;
 use App\Http\Controllers\Controller;
 use App\Models\Item;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ItemController extends Controller
 {
@@ -30,7 +31,11 @@ class ItemController extends Controller
      */
     public function create()
     {
-        //
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
+        return "Item_create";
     }
 
     /**
@@ -91,6 +96,11 @@ class ItemController extends Controller
     public function edit(Item $item)
     {
         //
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
+        return "Item_edit";
     }
 
     /**
@@ -130,6 +140,11 @@ class ItemController extends Controller
      */
     public function destroy(Item $item)
     {
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
         $item->delete();
+        return "Item_destroy";
     }
 }
