+++ app/Http/Controllers/api/v1/LoanRequestController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\api\v1;
 use App\Http\Controllers\Controller;
 use App\Models\LoanRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class LoanRequestController extends Controller
 {
@@ -90,6 +91,11 @@ class LoanRequestController extends Controller
     public function edit(LoanRequest $loanRequest)
     {
         //
+        if(!Gate::allows(['admin', 'operator'])){
+            abort(403);
+        }
+
+        return "LoanRequest_edit";
     }
 
     /**
@@ -128,6 +134,11 @@ class LoanRequestController extends Controller
      */
     public function destroy(LoanRequest $loanRequest)
     {
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
         $loanRequest->delete();
+        return "LoanRequest_destroy";
     }
 }
